+++ src/test/php/net/xp_framework/unittest/text/doclet/Namespaced.class.php
@@ -1,8 +1,0 @@
-<?php namespace net\xp_framework\unittest\text\doclet;
-
-/**
- * Class Namespaced (used in ClassDocTest)
- *
- */
-class Namespaced extends \Object {
-}